Update PICTURE examples at top of file with new function names
Some of the commented-out PICTURE forms at the top of the file
pass symbols for the FN argument that are not bound to a
function. Probably, the names of these functions changed over
time. If you try to eval any of these forms, you'll get a slightly
confusing error, since the error about the unbound function is
swallowed by the IGNORE-ERRORS in PARAMETRIC-PATH. PARAMETRIC-PATH
winds up returning '() in this case, which means that P0-P7 are
similarly bound to '() in SHADED-ANNULUS, and finally we wind up
choking on the NIL in CAREFUL-ABS, with a backtrace like so:

  0: (REALPART NIL)
  1: (CAREFUL-ABS NIL)
  2: (FAR-OUT NIL)
  3: (SPLICE NIL NIL)
  4: (SHADED-ANNULUS ...)

Update these comments with a best-guess at the corrected function.

minus-over-plus -> one-minus-over-one-plus
sqrt-one-minus-sq -> sqrt-square-minus-one
series -> sqrt-one-plus-square

The first two seem like reasonably safe bets; that last one is less
certain.